@@ -7,8 +7,7 @@ import 'animation_slide_up_down_fade.dart';
 class CarouselSlide3 extends StatefulWidget {
   final int slideDuration;
 
-  const CarouselSlide3({Key? key, required this.slideDuration})
-      : super(key: key);
+  const CarouselSlide3({super.key, required this.slideDuration});
 
   @override
   _CarouselSlide3State createState() => _CarouselSlide3State();
